Revisi Schema DB & CRUD Movie

+++ app/Http/Controllers/GenreFilmController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Alert;
 use App\Models\GenreFilm;
+use App\Models\Movie;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Validator;
@@ -128,6 +129,11 @@ class GenreFilmController extends Controller
      */
     public function destroy($id)
     {
+        if (Movie::where('id_genre', $id)->count() > 0) {
+            Alert::error('Fail', 'Gagal Menghapus Genre, Masih ada movie dengan genre ini!');
+            return redirect()->route('genre-film.index');
+        }
+
         $genre_film = GenreFilm::findOrFail($id);
         $genre_film->delete();
         Alert::success('Done', 'Data Berhasil Dihapus!')->autoClose(3000);
